requirements.txt
coverage==4.5.2
Django==2.1.4
django-nose==1.4.6
djangorestframework==3.9.0
freeze==1.0.10
lxml==4.2.5
nose==1.3.7
packtools==2.3.8
picles.plumber==0.11
Pillow==5.3.0
pytz==2018.7
six==1.11.0
